+++ hydra-node/src/Hydra/Chain/Blockfrost.hs
@@ -7,13 +7,14 @@ import Control.Concurrent.Class.MonadSTM (putTMVar, readTQueue, readTVarIO, take
 import Control.Exception (IOException)
 import Control.Monad.Catch (Handler (Handler))
 import Control.Monad.Catch qualified as Catch
-import Control.Retry (RetryPolicyM, RetryStatus (..), capDelay, constantDelay, fullJitterBackoff, limitRetries, recovering, retrying)
+import Control.Retry (RetryPolicyM, RetryStatus (..), constantDelay, recovering, retrying)
 import Data.ByteString.Base16 qualified as Base16
 import Data.Text qualified as T
 import Hydra.Cardano.Api (
   BlockHeader (..),
   ChainPoint (..),
   Hash,
+  NetworkId,
   SlotNo (..),
   Tx,
   deserialiseFromCBOR,
@@ -24,7 +25,7 @@ import Hydra.Cardano.Api (
  )
 import Hydra.Chain (ChainComponent, ChainStateHistory, PostTxError (..), prefixOf)
 import Hydra.Chain.Backend (ChainBackend (..))
-import Hydra.Chain.Blockfrost.Client (APIBlockfrostError (..), isRetryable)
+import Hydra.Chain.Blockfrost.Client (APIBlockfrostError (..), blockfrostRetryPolicy, isRetryable, submissionRetryPolicy)
 import Hydra.Chain.Blockfrost.Client qualified as Blockfrost
 import Hydra.Chain.CardanoClient qualified as CardanoClient
 import Hydra.Chain.Direct.Handlers (
@@ -35,12 +36,12 @@ import Hydra.Chain.Direct.Handlers (
   newLocalChainState,
  )
 import Hydra.Chain.Direct.State (ChainContext)
-import Hydra.Chain.Direct.TimeHandle (queryTimeHandle)
+import Hydra.Chain.Direct.TimeHandle (newCachedTimeHandle, queryTimeHandle)
 import Hydra.Chain.Direct.Wallet (TinyWallet (..))
 import Hydra.Logging (Tracer, traceWith)
 import Hydra.Options (BlockfrostOptions (..), CardanoChainConfig (..))
 
-newtype BlockfrostBackend a = BlockfrostBackend (ReaderT BlockfrostOptions IO a)
+newtype BlockfrostBackend a = BlockfrostBackend (ReaderT BlockfrostEnv IO a)
   deriving newtype
     ( Functor
     , Applicative
@@ -50,70 +51,89 @@ newtype BlockfrostBackend a = BlockfrostBackend (ReaderT BlockfrostOptions IO a)
     , MonadCatch
     )
 
+data BlockfrostEnv = BlockfrostEnv
+  { project :: Blockfrost.Project
+  , genesisVar :: TVar IO (Maybe Blockfrost.Genesis)
+  }
+
+newBlockfrostEnv :: BlockfrostOptions -> IO BlockfrostEnv
+newBlockfrostEnv BlockfrostOptions{projectPath} =
+  BlockfrostEnv <$> Blockfrost.projectFromFile projectPath <*> newLabelledTVarIO "blockfrost-genesis-cache" Nothing
+
 runBlockfrostBackend :: BlockfrostOptions -> BlockfrostBackend a -> IO a
-runBlockfrostBackend opts (BlockfrostBackend m) = runReaderT m opts
+runBlockfrostBackend opts action = newBlockfrostEnv opts >>= (`runBlockfrostBackendWith` action)
+
+runBlockfrostBackendWith :: BlockfrostEnv -> BlockfrostBackend a -> IO a
+runBlockfrostBackendWith env (BlockfrostBackend m) = runReaderT m env
+
+-- | Return the cached value or run the action once and store its result.
+-- Concurrent callers may run the action more than once (last write wins),
+-- which is safe for immutable values.
+memoizeIO :: TVar IO (Maybe a) -> IO a -> IO a
+memoizeIO var action = do
+  memoized <- readTVarIO var
+  case memoized of
+    Nothing -> do
+      result <- action
+      atomically $ writeTVar var (Just result)
+      pure result
+    Just d -> pure d
+
+cachedGenesis :: BlockfrostBackend Blockfrost.Genesis
+cachedGenesis = BlockfrostBackend $ do
+  BlockfrostEnv{project, genesisVar} <- ask
+  liftIO $ memoizeIO genesisVar $ Blockfrost.runBlockfrostM project Blockfrost.queryGenesisParameters
 
 instance ChainBackend BlockfrostBackend where
-  queryGenesisParameters = withProject $ \_ prj ->
-    Blockfrost.toCardanoGenesisParameters <$> Blockfrost.runBlockfrostM prj Blockfrost.queryGenesisParameters
+  queryGenesisParameters = Blockfrost.toCardanoGenesisParameters <$> cachedGenesis
 
-  queryScriptRegistry txIds = withProject $ \opts prj ->
-    Blockfrost.runBlockfrostM prj $ Blockfrost.queryScriptRegistry opts txIds
+  queryScriptRegistry txIds = withNetworkId $ \networkId ->
+    Blockfrost.queryScriptRegistry networkId txIds
 
-  queryNetworkId = withProject $ \_ prj -> do
-    -- TODO: This calls to queryGenesisParameters again, but we only need the network magic
-    Blockfrost.Genesis{_genesisNetworkMagic} <- Blockfrost.runBlockfrostM prj Blockfrost.queryGenesisParameters
-    pure $ Blockfrost.toCardanoNetworkId _genesisNetworkMagic
+  queryNetworkId = Blockfrost.toCardanoNetworkId . Blockfrost._genesisNetworkMagic <$> cachedGenesis
 
-  queryTip = withProject $ \_ prj ->
-    Blockfrost.runBlockfrostM prj Blockfrost.queryTip
+  queryTip = runQuery Blockfrost.queryTip
 
-  queryUTxO addresses = withProject $ \_ prj -> do
-    Blockfrost.Genesis{_genesisNetworkMagic} <-
-      Blockfrost.runBlockfrostM prj Blockfrost.queryGenesisParameters
-    let networkId = Blockfrost.toCardanoNetworkId _genesisNetworkMagic
-    Blockfrost.runBlockfrostM prj $ Blockfrost.queryUTxO networkId addresses
+  queryUTxO addresses = withNetworkId $ \networkId ->
+    Blockfrost.queryUTxO networkId addresses
 
-  queryUTxOByTxIn txins = withProject $ \opts prj -> do
-    Blockfrost.Genesis{_genesisNetworkMagic} <-
-      Blockfrost.runBlockfrostM prj Blockfrost.queryGenesisParameters
-    let networkId = Blockfrost.toCardanoNetworkId _genesisNetworkMagic
-    Blockfrost.runBlockfrostM prj $ Blockfrost.queryUTxOByTxIn opts networkId txins
+  queryUTxOByTxIn txins = withNetworkId $ \networkId ->
+    Blockfrost.queryUTxOByTxIn networkId txins
 
-  queryEraHistory _ = withProject $ \_ prj ->
-    Blockfrost.runBlockfrostM prj Blockfrost.queryEraHistory
+  queryEraHistory _ = runQuery Blockfrost.queryEraHistory
 
-  querySystemStart _ = withProject $ \_ prj ->
-    Blockfrost.runBlockfrostM prj Blockfrost.querySystemStart
+  querySystemStart _ = Blockfrost.toCardanoSystemStart <$> cachedGenesis
 
-  queryProtocolParameters _ = withProject $ \_ prj ->
-    Blockfrost.runBlockfrostM prj Blockfrost.queryProtocolParameters
+  queryProtocolParameters _ = runQuery Blockfrost.queryProtocolParameters
 
-  queryStakePools _ = withProject $ \_ prj ->
-    Blockfrost.runBlockfrostM prj Blockfrost.queryStakePools
+  queryStakePools _ = runQuery Blockfrost.queryStakePools
 
-  queryUTxOFor _ vk = withProject $ \_ prj ->
-    Blockfrost.runBlockfrostM prj $ Blockfrost.queryUTxOFor vk
+  queryUTxOFor _ vk = withNetworkId $ \networkId ->
+    Blockfrost.queryUTxOFor networkId vk
 
-  submitTransaction tx = withProject $ \_ prj ->
-    void $ Blockfrost.runBlockfrostM prj $ Blockfrost.submitTransaction tx
+  submitTransaction tx = void . runQuery $ Blockfrost.submitTransaction tx
 
-  awaitTransaction tx vk = withProject $ \opts prj ->
-    Blockfrost.runBlockfrostM prj $ Blockfrost.awaitTransaction opts tx vk
+  awaitTransaction tx vk = withNetworkId $ \networkId ->
+    Blockfrost.awaitTransaction networkId tx vk
 
-  getBlockTime = withProject $ \_ prj -> do
-    Blockfrost.Genesis{_genesisActiveSlotsCoefficient, _genesisSlotLength} <-
-      Blockfrost.runBlockfrostM prj Blockfrost.queryGenesisParameters
+  getBlockTime = do
+    Blockfrost.Genesis{_genesisActiveSlotsCoefficient, _genesisSlotLength} <- cachedGenesis
     pure $ CardanoClient.computeBlockTime (fromInteger _genesisSlotLength) _genesisActiveSlotsCoefficient
 
-withProject :: (BlockfrostOptions -> Blockfrost.Project -> IO a) -> BlockfrostBackend a
-withProject f = BlockfrostBackend $ do
-  opts@BlockfrostOptions{projectPath} <- ask
-  prj <- liftIO $ Blockfrost.projectFromFile projectPath
-  liftIO $ f opts prj
+-- | Run a Blockfrost client action with the backend's project.
+runQuery :: Blockfrost.BlockfrostClientT IO a -> BlockfrostBackend a
+runQuery action = BlockfrostBackend $ do
+  BlockfrostEnv{project} <- ask
+  liftIO $ Blockfrost.runBlockfrostM project action
+
+-- | Run a Blockfrost client action that needs the (cached) network id.
+withNetworkId :: (NetworkId -> Blockfrost.BlockfrostClientT IO a) -> BlockfrostBackend a
+withNetworkId action = do
+  networkId <- queryNetworkId
+  runQuery (action networkId)
 
 withBlockfrostChain ::
-  BlockfrostOptions ->
+  BlockfrostEnv ->
   Tracer IO CardanoChainLog ->
   CardanoChainConfig ->
   ChainContext ->
@@ -121,7 +141,7 @@ withBlockfrostChain ::
   -- | Chain state loaded from persistence.
   ChainStateHistory Tx ->
   ChainComponent Tx IO a
-withBlockfrostChain opts tracer config ctx wallet chainStateHistory callback action = do
+withBlockfrostChain env tracer config ctx wallet chainStateHistory callback action = do
   -- Known points on chain as loaded from persistence.
   let persistedPoints = prefixOf chainStateHistory
 
@@ -137,10 +157,11 @@ withBlockfrostChain opts tracer config ctx wallet chainStateHistory callback act
   -- Use the tip if we would otherwise start at the genesis (it can't be a good choice).
   prefix <-
     case head startFromPrefix of
-      ChainPointAtGenesis -> runBlockfrostBackend opts queryTip <&> (:| [])
+      ChainPointAtGenesis -> runBlockfrostBackendWith env queryTip <&> (:| [])
       _ -> pure startFromPrefix
 
-  let getTimeHandle = runBlockfrostBackend opts queryTimeHandle
+  let getTimeHandle = runBlockfrostBackendWith env queryTimeHandle
+  cachedTimeHandle <- newCachedTimeHandle (runBlockfrostBackendWith env)
   localChainState <- newLocalChainState chainStateHistory
   queue <- newLabelledTQueueIO "blockfrost-chain-queue"
   let chainHandle =
@@ -152,20 +173,20 @@ withBlockfrostChain opts tracer config ctx wallet chainStateHistory callback act
           localChainState
           (submitTx queue)
 
-  let handler = chainSyncHandler tracer callback getTimeHandle ctx localChainState
+  let handler = chainSyncHandler tracer callback cachedTimeHandle ctx localChainState
+  let getGenesis = liftIO (runBlockfrostBackendWith env cachedGenesis)
   res <-
     raceLabelled
       ( "blockfrost-chain-connection"
       , handle onIOException $ do
-          prj <- Blockfrost.projectFromFile projectPath
-          blockfrostChain tracer queue prj prefix handler wallet
+          blockfrostChain tracer queue project getGenesis prefix handler wallet
       )
       ("blockfrost-chain-handle", action chainHandle)
   case res of
     Left () -> error "'connectTo' cannot terminate but did?"
     Right a -> pure a
  where
-  BlockfrostOptions{projectPath} = opts
+  BlockfrostEnv{project} = env
   CardanoChainConfig{startChainFrom} = config
 
   submitTx :: TQueue IO (Tx, TMVar IO (Maybe (PostTxError Tx))) -> Tx -> IO ()
@@ -196,14 +217,15 @@ blockfrostChain ::
   Tracer m CardanoChainLog ->
   TQueue m (Tx, TMVar m (Maybe (PostTxError Tx))) ->
   Blockfrost.Project ->
+  m Blockfrost.Genesis ->
   NonEmpty ChainPoint ->
   ChainSyncHandler m ->
   TinyWallet m ->
   m ()
-blockfrostChain tracer queue prj prefix handler wallet = do
+blockfrostChain tracer queue prj getGenesis prefix handler wallet = do
   forever $
     raceLabelled_
-      ("blockfrost-chain-follow", blockfrostChainFollow tracer prj prefix handler wallet)
+      ("blockfrost-chain-follow", blockfrostChainFollow tracer prj getGenesis prefix handler wallet)
       ("blockfrost-submission", blockfrostSubmissionClient tracer (submitViaBlockfrost prj) queue)
 
 blockfrostChainFollow ::
@@ -211,17 +233,17 @@ blockfrostChainFollow ::
   (MonadIO m, MonadFail m, MonadCatch m, MonadDelay m, MonadLabelledSTM m, Catch.MonadMask m) =>
   Tracer m CardanoChainLog ->
   Blockfrost.Project ->
+  m Blockfrost.Genesis ->
   NonEmpty ChainPoint ->
   ChainSyncHandler m ->
   TinyWallet m ->
   m ()
-blockfrostChainFollow tracer prj prefix handler wallet = do
+blockfrostChainFollow tracer prj getGenesis prefix handler wallet = do
   -- Genesis query and start point resolution are wrapped in retry to survive
   -- transient HTTP errors (e.g. 403 rate limiting, connection resets).
   (blockTime, stateTVar) <-
     retryOnBlockfrostError tracer blockfrostRetryPolicy $ \_ -> do
-      Blockfrost.Genesis{_genesisSlotLength, _genesisActiveSlotsCoefficient} <-
-        Blockfrost.runBlockfrostM prj Blockfrost.getLedgerGenesis
+      Blockfrost.Genesis{_genesisSlotLength, _genesisActiveSlotsCoefficient} <- getGenesis
       let blockTime :: Double = realToFrac _genesisSlotLength / realToFrac _genesisActiveSlotsCoefficient
       -- Start from the latest point and fall back to older ones (best effort)
       -- If none of them can be resolved, we fall back to the tip of the chain.
@@ -229,11 +251,15 @@ blockfrostChainFollow tracer prj prefix handler wallet = do
       stateTVar <- newLabelledTVarIO "blockfrost-chain-state" blockHash
       pure (blockTime, stateTVar)
 
-  void $
-    retrying (retryPolicy blockTime) shouldRetry $ \_ -> do
-      pollForNewBlocks blockTime stateTVar
-        `catch` \(ex :: APIBlockfrostError) ->
-          pure $ Left ex
+  either throwIO pure
+    =<< retrying
+      (retryPolicy blockTime)
+      shouldRetry
+      ( \_ -> do
+          pollForNewBlocks blockTime stateTVar
+            `catch` \(ex :: APIBlockfrostError) ->
+              pure $ Left ex
+      )
  where
   shouldRetry :: x -> Either APIBlockfrostError a -> m Bool
   shouldRetry _ = \case
@@ -349,7 +375,7 @@ blockfrostSubmissionClient tracer submit queue = bfClient
 submitViaBlockfrost :: MonadIO m => Blockfrost.Project -> Tx -> m (Either Text Blockfrost.TxHash)
 submitViaBlockfrost prj tx =
   liftIO $
-    (Right <$> Blockfrost.runBlockfrostM prj (Blockfrost.submitTransaction tx))
+    (Right <$> Blockfrost.runBlockfrostMWith submissionRetryPolicy prj (Blockfrost.submitTransaction tx))
       `catch` (\(e :: APIBlockfrostError) -> pure . Left $ show e)
       `catch` (\(e :: IOException) -> pure . Left $ show e)
 
@@ -363,21 +389,10 @@ toChainPoint Blockfrost.Block{_blockSlot, _blockHash} =
   headerHash :: Hash BlockHeader
   headerHash = fromString . toString $ Blockfrost.unBlockHash _blockHash
 
--- * Retry logic
-
--- | Maximum number of retries for transient Blockfrost errors.
-maxRetries :: Int
-maxRetries = 10
-
 -- | Maximum number of blocks fetched per poll iteration (the Blockfrost page
 -- size limit).
 maxBlockBatch :: Int
 maxBlockBatch = 100
-
--- | Retry policy for transient Blockfrost errors: full-jitter exponential
--- backoff with 1s base, capped at 60s, at most 'maxRetries' retries.
-blockfrostRetryPolicy :: MonadIO m => RetryPolicyM m
-blockfrostRetryPolicy = capDelay 60_000_000 (fullJitterBackoff 1_000_000) <> limitRetries maxRetries
 
 retryOnBlockfrostError ::
   (MonadIO m, Catch.MonadMask m) =>